+++ src/service/future.rs
@@ -3,17 +3,22 @@
 use crate::{
     body::{box_body, BoxBody},
     response::IntoResponse,
+    routing::RouteFuture,
 };
 use bytes::Bytes;
 use futures_util::ready;
-use http::Response;
+use http::{
+    header::{HeaderValue, CONTENT_LENGTH},
+    Method, Request, Response,
+};
+use http_body::Empty;
 use pin_project_lite::pin_project;
 use std::{
     future::Future,
     pin::Pin,
     task::{Context, Poll},
 };
-use tower::BoxError;
+use tower::{BoxError, Service};
 
 pin_project! {
     /// Response future for [`HandleError`](super::HandleError).
@@ -48,5 +53,42 @@ where
                 }
             }
         }
+    }
+}
+
+pin_project! {
+    /// Response future for [`crate::handler::OnMethod`] and
+    /// [`crate::service::OnMethod`].
+    pub struct OnMethodResponseFuture<S, F, B>
+    where
+        S: Service<Request<B>>,
+        F: Service<Request<B>>
+    {
+        #[pin]
+        pub(crate) f: RouteFuture<S, F, B>,
+        pub(crate) req_method: Method,
+    }
+}
+
+impl<S, F, B> Future for OnMethodResponseFuture<S, F, B>
+where
+    S: Service<Request<B>, Response = Response<BoxBody>>,
+    F: Service<Request<B>, Response = Response<BoxBody>, Error = S::Error>,
+{
+    type Output = Result<Response<BoxBody>, S::Error>;
+
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let this = self.project();
+        let mut res: Response<_> = ready!(this.f.poll(cx)?);
+
+        // HEAD must not contain a body according to
+        // https://httpwg.org/specs/rfc7231.html#HEAD
+        if *this.req_method == Method::HEAD {
+            res.headers_mut()
+                .insert(CONTENT_LENGTH, HeaderValue::from_static("0"));
+            res = res.map(|_| box_body(Empty::new()));
+        }
+
+        Poll::Ready(Ok(res))
     }
 }